Switch Vim plugin manager from NeoBundle to Vim-Plug.
Switched Vim plugin manager from NeoBundle to Vim-Plug.  NeoBundle is no
longer maintained, and thus this seemed like prudent thing to do, to
attempt to address a bug in one of the plugins.

From what I can tell so far, too, Vim-Plug is a bit better than
NeoBundle, so I'm glad to make the switch, regardless!